Tutorials: Change directory structure
In preparation to pull in lilyglyphs and musicexamples I reorder directories.
This commit moves /documentation/tutorials to the root directory.
Tutorials are a main subproject and not part of the documentation

Additionally I create subdirectories for tutorial authors to make the structure more clear when there (hopefully) will be more than one tutorial for an author