Rename ulPlainTextWorkflows to plaintextworkflows-draft
After the Karlsruhe presentation I will completely review the text.
The first version has been written as a draft and as a collection of materials.

So I re-label the existing file as a draft and set up a new document.
While filling up this document (based on the outline of the Karlsruhe script)
I will move material from the draft to the new document and leave the remaining
stuff to be used in further texts