plaintext-de: Remove "openLilyLib Tutorials" dependencies
This version is an independent document.
Especially it's in German, and I didn't see a way to modify the main document
to support languages.
And I want to use my non-free fonts for that specific purpose.
So I keep the file in the tutorials directory but make it (partially) independent