agOrchestralScores: first review from Urs
I've mostly reviewed through asking questions in the red
\todo{} items.
I've mainly concentrated on structural elements, taking the
(natural) perspective of a reader who has to learn all that.
In some cases I have done little linguistic changes  but in
general I didn't care for the language so far.

The most important topics at the moment seem to be
- missing explanations (mostly short mentioning is enough)
- material that could be dropped
- explanations that should be moved somewhere else
  to make the structure clearer